Admin: Move dashboard blocks to separate admin modules
Move dashboard blocks for `OrderModule` and `ContactModules` to separate
admin modules `SalesDashboardModule` and `CustomerDashboardModule` so
permission group access can be configured for those specific modules.

Refs SHUUP-2973
