+++ doc/api/shuup.admin.modules.contacts.rst
@@ -8,18 +8,6 @@ Subpackages
 
     shuup.admin.modules.contacts.views
 
-Submodules
-----------
-
-shuup.admin.modules.contacts.dashboard module
----------------------------------------------
-
-.. automodule:: shuup.admin.modules.contacts.dashboard
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 Module contents
 ---------------
