+++ doc/api/shuup.admin.modules.customers_dashboard.rst
@@ -1,0 +1,22 @@
+shuup.admin.modules.customers_dashboard package
+===============================================
+
+Submodules
+----------
+
+shuup.admin.modules.customers_dashboard.dashboard module
+--------------------------------------------------------
+
+.. automodule:: shuup.admin.modules.customers_dashboard.dashboard
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: shuup.admin.modules.customers_dashboard
+    :members:
+    :undoc-members:
+    :show-inheritance:
