+++ doc/api/shuup.admin.modules.sales_dashboard.rst
@@ -1,0 +1,22 @@
+shuup.admin.modules.sales_dashboard package
+===========================================
+
+Submodules
+----------
+
+shuup.admin.modules.sales_dashboard.dashboard module
+----------------------------------------------------
+
+.. automodule:: shuup.admin.modules.sales_dashboard.dashboard
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: shuup.admin.modules.sales_dashboard
+    :members:
+    :undoc-members:
+    :show-inheritance:
