+++ shuup/admin/__init__.py
@@ -25,6 +25,7 @@ class ShuupAdminAppConfig(AppConfig):
             "shuup.admin.modules.categories:CategoryModule",
             "shuup.admin.modules.contacts:ContactModule",
             "shuup.admin.modules.contact_groups:ContactGroupModule",
+            "shuup.admin.modules.customers_dashboard:CustomersDashboardModule",
             "shuup.admin.modules.permission_groups:PermissionGroupModule",
             "shuup.admin.modules.users:UserModule",
             "shuup.admin.modules.service_providers:ServiceProviderModule",
@@ -32,6 +33,7 @@ class ShuupAdminAppConfig(AppConfig):
             "shuup.admin.modules.services:ShippingMethodModule",
             "shuup.admin.modules.attributes:AttributeModule",
             "shuup.admin.modules.sales_units:SalesUnitModule",
+            "shuup.admin.modules.sales_dashboard:SalesDashboardModule",
             "shuup.admin.modules.shops:ShopModule",
             "shuup.admin.modules.demo:DemoModule",
             "shuup.admin.modules.manufacturers:ManufacturerModule",
