+++ shuup/admin/locale/en/LC_MESSAGES/djangojs.po
@@ -8,14 +8,14 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-07-07 22:34+0000\n"
+"POT-Creation-Date: 2016-07-12 00:44+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: en <LL@li.org>\n"
+"Language: en\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
-"Language: en\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
 "Generated-By: Babel 2.1.1\n"
 
@@ -31,18 +31,21 @@ msgstr ""
 msgid "New file name?"
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "Are you sure you want to delete the file %s?"
 msgstr ""
 
 msgid "New folder name?"
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "Are you sure you want to delete the %s folder?"
 msgstr ""
 
 msgid "New folder"
+msgstr ""
+
+msgid "current"
 msgstr ""
 
 msgid "Oldest first"
@@ -92,7 +95,7 @@ msgstr ""
 msgid "Delete folder"
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "Order %s created."
 msgstr ""
 
@@ -277,7 +280,7 @@ msgid ""
 "customer first."
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "All prices are in %s currency."
 msgstr ""
 
@@ -332,7 +335,7 @@ msgstr ""
 msgid "Identifier"
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "Value %s Name"
 msgstr ""
 
@@ -378,20 +381,20 @@ msgstr ""
 msgid "The results could not be loaded"
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "Please delete %s character"
 msgid_plural "Please delete %s characters"
 msgstr[0] ""
 msgstr[1] ""
 
-#, c-format
+#, javascript-format
 msgid "Please enter %s or more characters"
 msgstr ""
 
 msgid "Loading more results..."
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "You can only select %s item"
 msgid_plural "You can only select %s items"
 msgstr[0] ""
@@ -436,7 +439,7 @@ msgstr ""
 msgid "Next"
 msgstr ""
 
-#, c-format
+#, javascript-format
 msgid "Filter by %s"
 msgstr ""
